refactor(run): move e2e tests to separate testing directory (#1681)
Fixes #1674 

I had to create a module for the new /run/testing directory to ensure the e2e tests are only run once instead of in every supported Go version.

More discussion on this in the issue.
