+++ run/grpc-ping/go.mod
@@ -3,7 +3,6 @@ module github.com/GoogleCloudPlatform/golang-samples/run/grpc-ping
 go 1.13
 
 require (
-	github.com/GoogleCloudPlatform/golang-samples v0.0.0-20200901140421-b993d0229728
 	github.com/golang/protobuf v1.4.2
 	google.golang.org/api v0.31.0
 	google.golang.org/grpc v1.31.1
