+++ run/grpc-server-streaming/go.mod
@@ -3,7 +3,13 @@ module github.com/GoogleCloudPlatform/golang-samples/run/grpc-server-streaming
 go 1.13
 
 require (
-	github.com/GoogleCloudPlatform/golang-samples v0.0.0-20200901140421-b993d0229728
 	github.com/golang/protobuf v1.4.2
+	github.com/google/go-cmp v0.5.2 // indirect
+	golang.org/x/net v0.0.0-20200822124328-c89045814202 // indirect
+	golang.org/x/sys v0.0.0-20200803210538-64077c9b5642 // indirect
+	golang.org/x/text v0.3.3 // indirect
+	golang.org/x/xerrors v0.0.0-20200804184101-5ec99f83aff1 // indirect
+	google.golang.org/genproto v0.0.0-20200827165113-ac2560b5e952 // indirect
 	google.golang.org/grpc v1.31.1
+	google.golang.org/protobuf v1.25.0 // indirect
 )
