+++ run/logging-manual/go.mod
@@ -2,7 +2,4 @@ module github.com/GoogleCloudPlatform/golang-samples/run/logging-manual
 
 go 1.12
 
-require (
-	cloud.google.com/go v0.65.0
-	github.com/GoogleCloudPlatform/golang-samples v0.0.0-20200901140421-b993d0229728
-)
+require cloud.google.com/go v0.65.0
