+++ run/markdown-preview/editor/go.mod
@@ -3,7 +3,6 @@ module github.com/GoogleCloudPlatform/golang-samples/run/markdown-preview/editor
 go 1.13
 
 require (
-	github.com/GoogleCloudPlatform/golang-samples v0.0.0-20200901140421-b993d0229728
 	golang.org/x/oauth2 v0.0.0-20200107190931-bf48bf16ab8d
 	google.golang.org/api v0.31.0
 )
