+++ run/markdown-preview/renderer/go.mod
@@ -3,8 +3,9 @@ module github.com/GoogleCloudPlatform/golang-samples/run/markdown-preview/render
 go 1.13
 
 require (
-	github.com/GoogleCloudPlatform/golang-samples v0.0.0-20200901140421-b993d0229728
 	github.com/microcosm-cc/bluemonday v1.0.4
+	github.com/pmezard/go-difflib v1.0.0 // indirect
 	github.com/russross/blackfriday/v2 v2.0.1
 	github.com/shurcooL/sanitized_anchor_name v1.0.0 // indirect
+	golang.org/x/net v0.0.0-20200822124328-c89045814202 // indirect
 )
