+++ run/pubsub/go.mod
@@ -1,5 +1,3 @@
 module github.com/GoogleCloudPlatform/golang-samples/run/pubsub
 
 go 1.12
-
-require github.com/GoogleCloudPlatform/golang-samples v0.0.0-20200901140421-b993d0229728
